requirements.txt
streamlit
langchain
faiss-cpu
python-docx
openai
tiktoken
python-dotenv